Add(update-bootengine): Move code for running dracut out of ebuild.
Trying to do all this during install in a chroot is crazy enough, we
need to be able to actually make this semi-safe. The big new thing
from what was previously in the ebuild is that this script makes sure it
is in its own filesystem namespace and then mounts directories into the
chroot privately. That way we don't alter the view everything else has
of this tree and eliminates the need to cleanup afterwards.

After this the ebuild just needs to do this in post install:

    update-bootengine -m -c "${ROOT}"