fix(update-bootengine): Compare namespaces via mountinfo.
The /proc/<pid>/ns/mnt link only exists in recent kernels. As-is this
causes update-bootengine to enter an endless exec loop because the if
statement only compares the readlink output, ignoring the exit value.
As an alternative we can read the contents of mountinfo, even though the
mounts are the same they will have different mount ids.