feat(parse-squashfs): add in support for custom OEMs
PXE images are a bit tricky because you only have two pieces of
information to work with: the kernel and the initramfs. The eventual
plan is for the kernel and initramfs to be one file on CoreOS. This
leaves the cpio free for user customization. So, we will allow users to
put their own OEM customizations into this dir.

The way this will work is that if there is a /usr/share/oem in the
initramfs then we mount a tmpfs on $NEWROOT/usr/share/oem