fix(tests): Rework the test suite, make false positives harder.
The test suite verified nothing exploded but often it was possible for
the wrong things to happen such as falling through to the fallback when
kexec was supposed to do something. This should be a bit more robust
(maybe) but perhaps redoing this with a proper test framework some time
would be wise.