fix(gptprio): Fix fallback path when using systemd.
When using dracut with systemd, dracut's own block device mounting
scripts are disabled in favor of relying on systemd's built in
fstab/cmdline parsing functionality. This means the existing behavior of
just setting the root= environment variable does nothing. The simplest
fix for this is to simply move our module to be a mount hook and leave
the fallback root filesystem mounted at /sysroot.