feat(80usrgptprio) add ability for unique /etc/machine-id files
At first boot, we will create a new /etc/machine-id file to make the
machine "unique".  If the file is not there, or empty, or conaining
things that do not make up a valid machine-id file, we will delete the
offending file and recreate it.

Thanks to @marineam for the hint to use grep to make things simpler.