add(10usr-generator): Move usr=/dev/foo handling to a systemd generator
The dracut parse hooks may not run early enough for
systemd-fstab-generator to catch what gets written to fstab at that
point. This means that /usr may not get mounted before systemd reaches
the initrd-root-fs.target and starts to switch the new root. This
manifests itself as random hangs during boot or the machine-id not being
properly initialized on first boot. Creating our own generator that
works like systemd-fstab-generator should fix this issue.

It would be nice to move this functionality into systemd itself.