update-bootengine: always run ldconfig in chroots before dracut
Portage is not good about updating ld.so.cache when installing packages
to alternate ROOTs. To ensure we are not operating with an old or
missing cache always run ldconfig before dracut. Helpfully dracut does
not fail when it cannot find some shared libraries, allowing it to
create completely broken initrds.