systemd: don't initialize machine-id, do initialize ld.so.conf
systemd for a while now has properly initialized machine-id if the file
is missing or blank. Switch to letting it do that by removing or bogus
machine id if it is present.

systemd does not always set up ld.so.conf early enough, switch to doing
so in the initrd. This removes the need for one of our systemd patches.