Godeps: drop GoogleCloudPlatform/kubernetes/pkg/api/resource
The Quantity type it provided hasn't worked out, our use case is
storage-centric and wants to always be base 2.  We'd also like things to be
more flexible/user-friendly.

The Quantity type was unhappy if there were spaces separating the number from
the unit, or if "KiB" were used instead of "Ki", annoying.