config/storage: Partition dimensions integration
This takes a stab at the details of sizing and placing partitions.

In YAML we require human units like MB,MiB,GB,GiB etc.

In JSON we require plain integers assumed to be 512-byte sector units.

When unmarshalling YAML we make some adjustments for alignment, and other
things like keeping partitions out of the GPT area.  Partition starts are
aligned to 2048-sector boundaries, which is what sgdisk expects.

One may specify 0 for a partition start to leave sgdisk to place the
partition.  The same may be done with size, leaving sgdisk to size the
partition to fill the contiguously available space within the region it's
placed.