exec/stages: move file-writing into disks stage
In order to write the files, the filesystem must be mounted to a
temporary location. If this is run during the "files" stage, the root
filesystem will already be mounted, preventing an additional mount.
Moving this into the "disks" stage fixes that (though the stage names no
longer make sense).