providers/cmdline: allow omission of config URL
On CoreOS, the default OEM is PXE (which translates to the cmdline
provider). If the user isn't intending on using Ignition, no config URL
will be provided. This change allows Ignition to handle this case by
no-op'ing.

This particular approach is extremely hacky, but this will all be gutted
soon. We can live with this for now.