verity-generator: disable default dependencies
By default services depend on `basic.target` but this is problematic for
anything directly in the dependency tree for `initrd-switch-root.target`
as it means everything under `basic.target` will not be stopped prior to
the switch root. This leads to all sorts of surprising things, including
stalling boot because `systemd-udev-trigger.service` fails to run after
switch-root, leaving udev with incomplete state.