config: add support for appending configs
When a config is appended to another, each field is iterated, appending slices
recursively appending structs, and overwriting old values with the new for all
other types. The exception to this rule are the fields in Config.Ignition.
Version from the original config is preserved and Config is replaced with the
newer config.