util: fix http client's handling of status codes
The body should only be returned if the response status code is a 200.
Otherwise, if the response status code is on the list of acceptable codes, an
empty body is returned. This will prevent EC2's 404 documents from being
interpretted as a valid config. Instead an empty document will be used.