setup-root: fix ordering with /sysroot/usr mount
First off, the script had a little bit of forgotten logic is left over
the read-only root and kexec days when things were much weirder and
allowing the setup to noop if /sysroot/usr wasn't mounted. This was
harmless until systemd 229 which includes the following change:

https://github.com/systemd/systemd/commit/104bc12fbc23c3ca852d9d389805c615cd590d01

So the fix is two fold: remove that bogus bit of logic so the setup
script explodes appropriately when /sysroot or /sysroot/usr isn't
mounted. Additionally fix the actual bug by ordering after all of the
possible paths under /sysroot that may be mounted.