oem: add support for default oem configs
This breaks OEM configs into two seperate types: the base config and the
default config. The base config is always applied and is appended to the
user's config. In the event that a user's config is empty or is a
cloud-config, the OEM's default config is used in its place and appended
to the OEM's base config.