providers/configdrive: add initial support for the openstack OEM
This uses the config drive's user data for the Ignition config.  It
should work with a nova configuration using either a disk or cdrom
device, as well as either vfat or iso9660.  The user data file can
currently be in either the OpenStack or EC2 latest API location, to
allow testing the provider on other platforms.