*: add support for allowing existing filesystems
This adds the 'useIfExists' flag to filesystem's 'mount' object. When
this flag is set, Ignition will not format the filesystem if it is
already of the correct type. This is useful if the user wants to persist
data across multiple runs of Ignition.

This new flag can be safely used in conjunction with the 'force' flag.
When both are set, the target filesystem is first checked to see if it's
the correct type. If so, the format operation is skipped. If not, the
'force' flag determines if the existing filesystem (if any) will be
overwritten.