dracut: Run ignition-files whenever /etc/machine-id doesn't exist
We currently use the disk GUID to determine when to run Ignition, but
then we reset the GUID whether or not it ran successfully. As a result,
rebooting after a failed Ignition (a fizzle?) will cause the system to
boot unprovisioned. We'd like to rerun Ignition on every boot until it
succeeds, but making disk-uuid@.service dependent on both
ignition-disks.service and ignition-files.service would cause disruptive
udev events after filesystems have been mounted.

Have ignition-files condition on the existence of /etc/machine-id instead
of the disk GUID. It still Requires ignition-disks, so will not run if
-disks fails. If -disks succeeds but -files fails (or /etc/machine-id is
manually removed), on the next boot -disks will be skipped but still
satisfy the dependency for -files.