dracut: Randomize disk GUID only after ignition-disks succeeds
We use the disk GUID to determine when to run ignition-disks, but we
currently reset the GUID whether or not ignition-disks was successful.
Rebooting after a failed run therefore causes the system to boot
unprovisioned.

Continue to run disk-uuid@.service only if coreos.randomize_disk_guid= is
specified on the kernel command line, but now also require that
ignition-disks either runs successfully or is skipped (because
coreos.first_boot=1 is not specified).