dracut: teach ignition-generator about kvm
Since systemd 233, a QEMU system using KVM has virtualization type
"kvm", as opposed to the previous "qemu".  (It was back to "kvm" at
some point several releases earlier.)  This sets the QEMU OEM ID
value for both now, so Ignition is run on QEMU regardless of the
systemd version or KVM support.