move storage.filesystem.mount.create.options up one level
The semantics of the filesystem create object are confusing and
undesirable, so this commit copies the options fields in this object up
one level into the mount object. Users may now specify these options in
either location, to do so in both is an error. If the old create object
is used, a warning is raised that this object is deprecated.