*: add support for reusing existing filesystems
This adds the 'wipeFilesystem' flag to filesystem's 'mount' object. When
this flag is false, Ignition will inspect the current filesystem's type,
and reuse it if it matches the desired state. If this flag is true, the
current filesystem will always be reformatted when Ignition is run.
This is useful if the user wants to persist data across multiple runs of
Ignition.

This new flag can be safely used in conjunction with the deprecated
'force' flag.  When both are set, the target filesystem is first checked
to see if it's the correct type. If so, the format operation is skipped.
If not, the 'force' flag determines if the existing filesystem (if any)
will be overwritten.