.travis.yml: drop support for go 1.5 and 1.6, and add support for go 1.8
This commit alters the travis configuration to stop running tests on go
1.5 and 1.6, and start running tests on go 1.8. This is necessary
because the `os/user.Group` struct used in `user_group_lookup.go` in the
prior commit was added to the stdlib in go 1.7.

Without further modification the tests will fail in go 1.8 due to a
gofmt change, so the changes necessary to fix this have been included as
well.