config/type/passwd: deprecate user create object
This commit moves all fields in the user create object up one level
directly into the user object. Using the create object and any of these
copied fields is an error. Using the create object generates a
deprecation warning.

When running the files stage, now ignition will first determine if the
described user exists. If the user exists, `usermod` will be used to
change the user to the desired state, and if the user does not exist
`useradd` will be used to create the user.