config/types/schema: autogenerated changes
The json schema generator ignition uses, schematyper, changed the name
of a type in the golang code with the schema change in the prior commit.
This is probably due to two concurrent schema changes in PRs, and
schematyper behaving differently with the two changes than when it only
had each one individually.

This commit should have no structural, semantic, or otherwise effects on
the project aside fom the name change.