internal/exec/stages/files: fix nil pointer dereference
Recent changes allowing node owners to be referenced by user/group names
changed the id field for the user and group to be a pointer, so that it
could be detected when it was left empty vs when it was specified to be
0. The simple example given to justify this was when a user provided a
config saying that the node should be owned by user "core" and uid "0",
if the id was not a pointer it would appear to Ignition as if they had
not provided a uid.

The logic dealing with setting the UID failed to properly handle nil
values here, and can cause nil pointer dereferences. This commit adds a
check and an assignment to fix this for both of these fields.

If a node's owner is unspecified, it will default to the root user and
group.