internal/resource/url: add S3 and IAM support
This commit modifies internal/resource/url to add support for the S3
schema. For authentication, Ignition will attempt to fetch IAM
credentials from the EC2 metadata service if Ignition is running in EC2.
If this fails or if Ignition is not running in EC2, it will fall back on
using the anonymous credentials.