config/types/schema: autogenerated changes
The json schema -> golang structs program we use, schematyper, for some
reason renamed a type in its output when I removed the aws object from
the schema. This commit includes that change and the resulting code
changes. This should not result in any change to Ignition, just the
naming of an internal type.