exec: propogate root parameter to baseconfig
If a value is passed to root other than the default of '/' then
the baseConfig will point to the given value rather than
'/sysroot'.

Also updates the RenderFile function from
'internal/exec/util/file.go' to be a member function of Util to
allow the given root parameter to be used rather than always using
'/sysroot'.