internal/providers/{ec2,gce}: ignore 404 errors
EC2 and GCE return an HTTP 404 error when Ignition attempts to fetch
its config and the user has not provided any user data. The recent
refactor of the fetching logic had a regression where this is no longer
handled, and the 404 would cause Ignition to fail. This commit makes
these providers ignore any 404 errors.