dracut: disable systemd-fsck@.service for /usr
On systems with verity enabled, systemd-fsck produces:

    systemd-fsck[392]: fsck.ext4: Operation not permitted while trying to open /dev/mapper/usr
    systemd-fsck[392]: You must have r/w access to the filesystem or be root
    systemd-fsck[392]: fsck failed with error code 8.
    systemd-fsck[392]: Ignoring error.

On systems with /usr formatted for verity, but with a bootloader too
old to enable it, systemd-fsck produces:

    systemd-fsck[369]: /dev/vda3 has unsupported feature(s): FEATURE_R24 FEATURE_R25 FEATURE_R26 FEATURE_R27 FEATURE_R28 FEATURE_R29 FEATURE_R30 FEATURE_R31
    systemd-fsck[369]: e2fsck: Get a newer version of e2fsck!
    systemd-fsck[369]: fsck failed with error code 8.
    systemd-fsck[369]: Ignoring error.

These errors are harmless but tend to confuse people. Add a generator,
based on 10usr-generator, that locates the /usr device and disables
systemd-fsck on it. As a consequence we no longer fsck /usr on the dev
image, but that seems like an acceptable tradeoff.