internal/exec/stages/disks: add --homehost any to mdadm creation
When the system hostname wasn't set in the initramfs, we were creating
RAID devices with a homehost of localhost. As a result, the /dev/md
symlink in the booted system could be either /dev/md/foo, if the device
was started before the hostname was set, or /dev/md/localhost:foo
otherwise.

Avoid this by creating arrays with the homehost set to any, effectively
disabling the homehost mechanism for those arrays. This ensures the
/dev/md symlink has a stable name, but causes the array not to be
treated as foreign if it is later connected to a different host.