config/config: parse bad configs for json errors
Fix an issue with config parsing where if a config's version was
indeterminable, it wasn't a cloud config, it wasn't empty, and it wasn't
a script, a generic "invalid version" message was emitted. The solution
is to run the unknown thing through ParseFromLatest to generate parse
errors.

Update the associated test to reflect different errors being returned in
some cases.