stages/disks: always settle after handling disks
Previously we only did a udevadm settle after creating filesystems;
this was added because creating btrfs filesystems was particularly good
at causing these races. Usually only settling after creating filesystems
is ok, because we typically create filesystems after partitioning. But
partitioning can create these kinds of races too, and if we do not
create filesystems we never settle and units dependent on Ignition and a
/dev symlink could find their link deleted out from under them.

This moves the settle up to happen after dealing with the disks stage in
entirety. It could be added as an ExecStartPost directive in the unit
file, but here we can be smart and skip it if the disks stage didn't do
anything.