exec/util/device_alias: retry evaluating symlinks
CreateDeviceAlias exists to create a stable symlink to be used so when
udev deletes and recreates the symlink it owns, Ignition wont fail.
However there is a narrow race creating the symlink: if udev creates
then deletes its symlink fast enough we wont have time to read the
symlink to create our own.

This adds retry logic to rapidly retry creating the symlink, in case we
lose the race, but still fail after a timeout.