*: add validate only flag
Developing validation logic either has a long test time due to needing
to build new images, or is dangerous due to needing to run Ignition
locally. This adds a --validate <ignfile> flag that allows Ignition to
just validate a config then exit.

It is implemented as a no-op stage and uses the file provider.