internal/resource: don't export client from the Fetcher
Management of the HttpClient should be an implementation detail of the
Fetcher. Don't export it, update code everywhere to not use it, and add
a function on the Fetcher to update the HTTP timeouts.

This results in a pattern change, where changing the HTTP timeouts now
doesn't require generating a new HttpClient and replacing the client in
the fetcher, and instead simply calling a function to mutate the
timeouts of the fetcher.