config/*: add systemd install section warning
This adds a warning to all config versions for an enabled unit with no
install section.

In general, any unit that has contents, is enabled, but has no install
section is useless.
It's unlikely a user intends to ever write such a unit, even if
technically the install section could be in a drop-in.

To avoid duplication, this also introduces a new 'shared' package under
'config'