build_blackbox_tests: use go build, use BIN_PATH
Recently this script started failing to place the stub binaries in
bin/${ARCH}, which would cause any tests using the passwd section of a
config to fail.

Update the script to build binaries with the same flags/env vars (except
ldflags since it doesn't do anything here) as is used in `build`.