Rework how we run Ignition on first boot
There was an extensive discussion of this on IRC.  The basic
idea is we have a stamp file in `/boot`, same as CL does.  However,
we use a GRUB *variable* `$coreos_firstboot` which is hardcoded in
the kernel arguments, but evaluates to the empty string if we're
not in the first boot.

This way we don't have to change the kernel arguments post-boot.

We previously had a systemd "quench" unit as a hardcoded Ignition
config here - I moved that to a regular installed unit that is
part of the package.