tests/register: run BB tests on all compatible config versions
This commit adds the ability to register multiple versions of a
single test. Copies of the test will be generated for every
version, inside the same major version, that is equal to or
greater than the specified ConfigMinVersion. For example, a test
with a 2.0.0 config will be registered as 4 identical tests.
A 2.0.0, 2.1.0, 2.2.0, and a 2.3.0-experimental test. This
improves test coverage by testing ignition's config backward
compatibility.

The minimum version a config supports is now specified
by the field ConfigMinVersion in the test struct. The
version in a BB test config should be $version and will
be updated before the test is run. This makes writing
tests less painful and aids in running  BB tests with
all compatible config versions.