tests/bb: BB test.config should no long have an explict version
The minimum version a config supports is now specified
by the field ConfigMinVersion in the test struct. The
version in a BB test config should be $version and will
be updated before the test is run. This makes writing
tests less painful and aids in running  BB tests with
all compatible config versions.