stages/files: relabel files for SELinux after boot
This patch teaches Ignition to detect if `restorecon` is in the root
filesystem. If that is the case and relabeling is enabled at compile
time, it will automatically track all files that it creates and add a
systemd unit file to relabel them all during early boot.

This is the least intrusive approach, though it comes with some caveats
as well. See [1] for an extensive discussion, but in a nutshell, there
is a bootstrapping problem -- using a relabeling service assumes that
the system will be able to successfully reach the point at which
relabeling can occur. OTOH, we can assume we're not dealing with a
completely unlabeled filesystem here. Only files created by Ignition
need relabeling, and those should be fine to relabel in time before the
rest of the system boots up.

[1] https://lists.fedoraproject.org/archives/list/devel@lists.fedoraproject.org/thread/CHCEGB2RUPHFCE4FVGIRO3CJYGNS75T7/