config/*/append*: remove unused code
Configs are first translated to the latest config, then appended. This
only the latest config needs to know how to handle appending. Remove all
other config appending code since it is just dead code.

n.b.: the experimental config is the latest but is translated to an
identical config in internal/config through an identity translation.
This internal config is the one which implements the appending.