tests/timeouts: fix backoff test race
Fix race where each config version test uses the same endpoint on the
backoff server. This causes the the time since the last request to be
reset by dualing tests. Fix it so each test hits a /$version endpoint.

Ultimately the tests should probably have a way of spinning up per-test
servers, but this works as a pretty good stopgap until a more proper
solution is written.