module-setup.sh: explicitly install qemu_fw_cfg
Make it explicit that we need the `qemu_fw_cfg` kernel module in the
initramfs. This works for Fedora right now before I think because we run
dracut with `--no-host-only` at compose time. (It's the default if
`initramfs-args` isn't in the manifest, as in FCOS, and in FAH we
explicitly have it in the manifest).

Closes: #22