ignition-disks.service: drop Requires=network.target
Mirror the same requirements as `ignition-files.service`. I.e. we don't
strictly *require* networking, but if there is networking, we'd rather
run after it so that we can pull down the config. See also #17.

(Anyway, I think we want `network-online.target` here instead of
`network.target` but I haven't tested that).