ignition: run files stage at original CL ordering
We previously messed around with changing the files service to
adapt to `ostree-prepare-root.service` running so late in the boot
process. Now that OSTree runs at more semantically appropriate times, we
can restore ordering here so that it runs right after
`initrd-root-fs.target`.

This also undoes the hacking around in `ignition-generator` to get the
services pulled in by `initrd-switch-root.service` instead of the more
appropriate principal target of the initrd (`initrd.target`).

Another nice thing about this is that we drop the explicit dependency
between `ostree-prepare-root.service` and `ignition-files` since they're
now synchronized through `initrd-root-fs.target`.

Closes: #20