dracut/30ignition: general hack cleanups
We're past the MVP phase now. Let's reflect that in our codebase by
dropping all the telltale signs of quick hacks and experimentations.
This makes the tree nicer to hack on as well.

Remember, all your code are belong to Git. We can always revive things
as needed as we get around to supporting Packet/DO, etc... Though those
units will need to be tweaked anyway.

Closes: #33