tests/positive/passwd: add test UseAuthorizedKeysFile
Add a blackbox test to verify that ~/.ssh/authorized_keys is written
when IGNITION_WRITE_AUTHORIZED_KEYS_FRAGMENT=false.

Also updates the existing test AddPasswdUsers to use the environment
variable IGNITION_WRITE_AUTHORIZED_KEYS_FRAGMENT=true so that
the test will pass consistently if the binary is built with a
non-default writeAuthorizedKeysFragment flag.