config/v3_0_exp/types: implement Key() for types
Implement a Key() function which returns a string uniquely identifying
members of a list to be used for duplication checking and config
merging.

Also implement MergedKeys() on types.Storage which returns a map of
struct members names to the common name that can refer to all them.
This allows checking lists that cannot have duplicates across lists.