*: default file.overwrite config option to false
Default file.overwrite in Ignition configs to false. This means that
if something exists already at file.path, give an error and do not
allow Ignition to finish. The exception is if file.append was
specified - then Ignition follows through with appending to a regular
file at file.path.

Previously, the existing file.path was deleted if the user user had
not set file.overwrite, before creating the new one.

Also update existing blackbox tests for the new behavior.

Resolves: #604