config/validate: validate no duplicates exist
Check that no duplicates exist in the config's lists. Check that
files/link/dirs don't have duplicates across lists

There is a special case here: lists of strings implicitly use the string
as its Key(). This is just a shortcut, we can drop it later if we need
to in favor of explicit Key() functions.