grub: support overriding network kcmdline args
This allows us to set network kcmdline args by setting the
ignition_network_kcmdline variable inside the ignition.firstboot
file that resides on the /boot/ filesystem and triggers ignition
run on boot.

This functionality will allow the coreos-installer to promote static
networking arguments that were used during the installer run to the
first boot of the real system. After that they are no longer needed
as ignition can configure networking as requested by the user.