dracut/30ignition: order ignition-complete.target before initrd.target
I think this was the missing piece in the issues described in #61.
Essentially, `ignition-complete.target` wasn't explicitly ordered wrt to
`initrd.target`, so its completion was racing against system's `isolate`
to `initrd-switch-root.target`.

Ordering ourselves before `initrd.target` (which is also the unit that
pulls us in through `initrd.target.requires`) ensures that the target
can only be reached successfully if we're successful. And if our target
fails, then we immediately trigger `emergency.target`.