dracut/30ignition: order ExecStop= units before initrd-switch-root.target
The way `isolate` works is interesting. When systemd does
`systemctl isolate initrd-switch-root.target`, it does two things: (1)
enqueue stop jobs for all the services not required for the new target,
and (2) enqueue start jobs for all the services required for the new
target.

However, there is no default ordering between the stop jobs and start
jobs. One would think that systemd by default would first stop all the
services, and only then starts the new services. That is not the case
(see [1] and [2]).

We need to explicitly tell systemd that we want our services to be
stopped first before starting the new services. The way to do this is to
add an explicit `Before=initrd-switch-root.target` in the units (see
[1]). Otherwise, the `ExecStop` in `coreos-mount-var.service` and
`ignition-mount.service` may not run before the switch root.

(Note systemd does serialize its state across the reexec, but it seems
like it doesn't correctly resume the "stop" jobs, presumably because it
can't access the systemd units anymore. Regardless, we definitely want
to umount everything before the switchroot anyway).

Closes: #70

[1] https://systemd-devel.freedesktop.narkive.com/wb1UNand
[2] https://github.com/systemd/systemd/issues/3436