tests: fix bb tests for size/start 0
Fix use of size/start instead of sizeMiB/startMiB. Fix partition table
to have an empty space before/after partition 9 instead of another
partition. This was causing the test to fail (negative test, ignition
incorrectly succeeded) since it was looking to verify partition 9 was as
large as possible/started at the begining on the largest available
block. With a partition before/after this was actually true, now with
blank space before/after it is not and Ignition fails corrects.