files/selinux: deglob user/group related paths
Deglob user and group related paths in ignition rather than relying on
restorecon to do it. We want to support systemd's
/run/systemd/relabel-extra.d which does not support globbing, so we
should do it here instead. This intentionally deglobs before adding to
the relabel list since only ignition internal paths should have globs.
Deglobbing when creating the unit could lead to unexpected behavior if a
user created a file with a wildcard in it.