engine: Write `/run/ignition.json` atomically
I just noticed this while reading the code; in practice, the write
should never fail, but we really want to write files atomically
as a general rule.  Particularly in this case as the code may
come back and try to read it again, and if we only did a partial
write we'd just continually fail to read it.